requirements.txt
# Vibing Letters - Python Dependencies

# Core libraries
numpy>=2.2.0
opencv-python>=4.12.0
pillow>=12.0.0
scipy>=1.16.0

# Morphing enhancements
perlin-noise>=1.12
easing-functions>=1.0.4

# Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: Additional utilities
matplotlib>=3.10.0
